creating Heap as class

+++ P1/randmst.cpp
@@ -1,9 +1,9 @@
 #include "graph.hpp"
 #include "heap.hpp"
 
-float mstSum(vector<vector<int>> mstEdges, vector<vector<double>> adjacencyMatrix) {
+double mstSum(vector<vector<int>> mstEdges, vector<vector<double>> adjacencyMatrix) {
     
-    float sum = 0;
+    double sum = 0;
 
     for (int i = 0; i < mstEdges.size(); i++) {
         sum += adjacencyMatrix[mstEdges[i][0]][mstEdges[i][1]];
@@ -25,7 +25,7 @@ int main(int argc, char* argv[])
 
     // Prim
     vector<int> heap = createHeap();
-    vector<float> d (n, inf);
+    vector<double> d (n, inf);
     vector<int> S (n, 0);
     vector<int> prev (n,0);
     vector<int> verticesInHeap (n, -1);
@@ -47,7 +47,7 @@ int main(int argc, char* argv[])
             
             if (v != u && S[v] == 0) {
 
-                float w = adjacencyMatrix[u][v];
+                double w = adjacencyMatrix[u][v];
                 if (d[v] > w) {
                     
                     d[v] = w;
